requirements.txt
google-genai
python-dotenv
pyopenssl
opencv-python
python-docx